Changes to be committed: 	renamed:    QQuantLib/utils/classical_finance.py -> QQuantLib/finance/classical_finance.py 	renamed:    benchmark/payoff_class.py -> QQuantLib/finance/payoff_class.py 	renamed:    benchmark/probability_class.py -> QQuantLib/finance/probability_class.py 	modified:   misc/notebooks/10_ApplicationTo_Finance_01_IntegralComputing.ipynb 	new file:   misc/notebooks/11_ApplicationTo_Finance_02_ClassicalFinance.ipynb
