+++ QQuantLib/finance/payoff_class.py
@@ -2,7 +2,7 @@
 Class definition for the PayOff
 """
 from functools import partial
-import QQuantLib.utils.classical_finance as cf
+import QQuantLib.finance.classical_finance as cf
 from QQuantLib.utils.utils import text_is_none
